Mv tests to test directory with an index file for exporting
