+++ client/Routes.js
@@ -5,16 +5,16 @@ import { Login, Signup } from './components/AuthForm';
 import Home from './components/Home';
 
 import { me } from './store';
-import Test1 from './components/Test1';
-import Test2 from './components/Test2';
-import Test3 from './components/Test3';
-import Test4 from './components/Test4';
-import Test5 from './components/Test5';
-import Test6 from './components/Test6';
-import Test7 from './components/Test7';
-import Test8 from './components/Test8';
-import Test9 from './components/Test9';
-import Test10 from './components/Test10';
+import Test1 from './components/Tests/Test1';
+import Test2 from './components/Tests/Test2';
+import Test3 from './components/Tests/Test3';
+import Test4 from './components/Tests/Test4';
+import Test5 from './components/Tests/Test5';
+import Test6 from './components/Tests/Test6';
+import Test7 from './components/Tests/Test7';
+import Test8 from './components/Tests/Test8';
+import Test9 from './components/Tests/Test9';
+import Test10 from './components/Tests/Test10';
 
 /**
  * COMPONENT
@@ -46,7 +46,7 @@ class Routes extends Component {
         ) : (
           <Switch>
             <Route path='/' exact component={Login} />
-            <Route path='/jest3' exact component={Test1} />
+            {/* <Route path='/jest3' exact component={Test1} /> */}
             <Route path='/login' component={Login} />
             <Route path='/signup' component={Signup} />
           </Switch>
