+++ client/components/Tests/Test8.js
@@ -9,7 +9,7 @@ import { javascript } from '@codemirror/lang-javascript';
 import { oneDark } from '@codemirror/theme-one-dark';
 import readOnlyRangesExtension from 'codemirror-readonly-ranges';
 import axios from 'axios';
-import { fetchPrompts } from '../store/prompts';
+import { fetchPrompts } from '../../store/prompts';
 import { autocompletion } from '@codemirror/autocomplete';
 import { connect } from 'react-redux';
 const { v4: uuidv4 } = require('uuid');
@@ -41,8 +41,8 @@ export const Editor = (props) => {
   const [response, setResponse] = useState('See your results here!');
   const { prompts } = props;
 
-  const templateTest = prompts[5]?.templateTest;
-  const narrative = prompts[5]?.narrative;
+  const templateTest = prompts[7]?.templateTest;
+  const narrative = prompts[7]?.narrative;
   const completions = [
     { label: 'toBe', type: 'keyword' },
     { label: 'expect', type: 'keyword' },
@@ -90,6 +90,25 @@ export const Editor = (props) => {
 
   useEffect(() => {
     turnOffCtrlS();
+    const addMarks = StateEffect.define();
+    const filterMarks = StateEffect.define();
+
+    const markField = StateField.define({
+      create() {
+        return Decoration.none;
+      },
+      update(value, tr) {
+        value = value.map(tr.changes);
+        for (let effect of tr.effects) {
+          if (effect.is(addMarks))
+            value = value.update({ add: effect.value, sort: true });
+          else if (effect.is(filterMarks))
+            value = value.update({ filter: effect.value });
+        }
+        return value;
+      },
+      provide: (f) => EditorView.decorations.from(f),
+    });
 
     const state = EditorState.create({
       doc: narrative || code2,
@@ -97,6 +116,7 @@ export const Editor = (props) => {
         basicSetup,
         oneDark,
         baseTheme,
+        markField,
         onUpdate2,
         javascript(),
         // removeIndentation(),
@@ -107,6 +127,13 @@ export const Editor = (props) => {
     const view2 = new EditorView({
       state,
       parent: editor2.current,
+    });
+    const strikeMark = Decoration.mark({
+      attributes: { style: 'color: white' },
+    });
+
+    view2.dispatch({
+      effects: addMarks.of([strikeMark.range(33, 1000)]),
     });
 
     const fetchStuff = async () => {
@@ -129,11 +156,11 @@ export const Editor = (props) => {
     return [
       {
         from: editor.doc.line(1).from,
-        to: editor.doc.line(5).to,
+        to: editor.doc.line(2).to,
       },
       {
-        from: editor.doc.line(7).from,
-        to: editor.doc.line(14).to,
+        from: editor.doc.line(4).from,
+        to: editor.doc.line(5).to,
       },
     ];
   };
@@ -180,8 +207,13 @@ export const Editor = (props) => {
     const strikeMark = Decoration.mark({
       attributes: { style: 'background: yellow' },
     });
+
     view.dispatch({
-      effects: addMarks.of([strikeMark.range(6, 19)]),
+      effects: addMarks.of([
+        strikeMark.range(11, 24),
+        strikeMark.range(50, 63),
+        strikeMark.range(95, 108),
+      ]),
     });
 
     const fetchStuff = async () => {
@@ -234,7 +266,7 @@ export const Editor = (props) => {
   return (
     <div className='p-5'>
       <div ref={editor2}></div>
-      <div className='p-5 font-bold'>{prompts[5]?.prompt}</div>
+      <div className='p-5 font-bold'>{prompts[7]?.prompt}</div>
       <div ref={editor}></div>
       <button className='m-5 bg-gray-400 p-1' onClick={onSubmit}>
         Evaluate Your Test
