+++ client/components/Tests/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import * as AllTests from './';
+
+const PaginatedTests = () => {
+  const Test1 = AllTests[0];
+  return <Test1></Test1>;
+};
+
+export default PaginatedTests;
